+++ CHANGELOG.md
@@ -1,13 +1,14 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
-## 3.1 Changes on 2021-04
+## 3.1 Changes on 2021-08
 - Amount is now optional on voids, collections and refunds.   If not specified, the remaining amount of the original transaction will be used.
-- Update Api-Version to 6.3
+- Update Api-Version to 6.6
 - Allow WebPaymentReference to be set on PaymentModel
 - Add AuthCode to initial receipt response
 - Add AuthCode, WebPaymentReference and Acquirer to response of Transactions.Get(receiptId)
 - New exemption flags ChallengeRequestIndicator and ScaExemption added to the ThreeDSecureTwo Model
+- Add PrimaryAccountDetails to RegisterCardModel and CheckCardModel
 
 ## 3.0 Changes on 2021-03-11
 - Update default base Urls with certificate pinning checks